Added: Scoring for both players. Working on getting the return-angle to change a bit depending on where on the paddle the ball hits. Not quite working yet.

+++ ball.h
@@ -5,8 +5,8 @@ typedef struct ball ball_t;
 struct ball {
 	int ballx;
 	int bally;
-	int speedx;
-	int speedy;
+	float speedx;
+	float speedy;
 	int radius;
 };
